+++ changelog.md
@@ -7,8 +7,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [4.1.0]
 
 * add ability to use CDN geolocation header for consent targeting
-* adhere to global privacy control header if geolocation is enabled
-* extend CookieConset cookie expiry from 2 to 12 months
+* add opt-out and do-not-sell popups for configured countries
+* adhere to global privacy control header
+* extend CookieConset cookie default expiry from 2 to 12 months
+* fix removal of cookies if consent is withdrawn
 
 ## [4.0.0]